+++ src/types/query.d.ts
@@ -14,6 +14,8 @@ export module query {
   export function Object(fields: ExprArg): Expr;
   export function Lambda(f: Lambda): Expr;
   export function Lambda(var_name: ExprArg, expr: ExprArg): Expr;
+  export function Call(ref: ExprArg, ...args: ExprArg[]): Expr;
+  export function Query(lambda: Lambda): Expr;
   export function Map(collection: ExprArg, lambda_expr: Lambda): Expr;
   export function Foreach(collection: ExprArg, lambda_expr: Lambda): Expr;
   export function Filter(collection: ExprArg, lambda_expr: Lambda): Expr;
@@ -37,6 +39,7 @@ export module query {
   export function CreateDatabase(params: ExprArg): Expr;
   export function CreateIndex(params: ExprArg): Expr;
   export function CreateKey(params: ExprArg): Expr;
+  export function CreateFunction(params: ExprArg): Expr;
 
   export function Match(index: ExprArg, ...terms: ExprArg[]): Expr;
   export function Union(...sets: ExprArg[]): Expr;